Making all data and methods associated with navigation of the bus as a dependency and injection it to bus agent - better encapsulation
