+++ main.py
@@ -10,6 +10,7 @@ from agents.director import Director
 from agents.bus import Bus
 from credentials import *
 from bus_lines.bus_line import BusLine
+from bus.bus_navigator import BusNavigator
 
 import time
 
@@ -17,14 +18,23 @@ import time
 if __name__ == "__main__":
     bus_line_414 = BusLine(414, 18400)
     director = Director(DIRECTOR, DIRECTOR_PASSWD)
+
     bus1 = Bus(BUS1, BUS1_PASSWD)
-    bus1.add_line(bus_line_414)
+    bus_navigator_1 = BusNavigator(bus_line_414)
+    bus1.add_bus_navigator(bus_navigator_1)
+
     bus2 = Bus(BUS2, BUS2_PASSWD)
-    bus2.add_line(bus_line_414)
+    bus_navigator_2 = BusNavigator(bus_line_414)
+    bus2.add_bus_navigator(bus_navigator_2)
+
     bus3 = Bus(BUS3, BUS3_PASSWD)
-    bus3.add_line(bus_line_414)
+    bus_navigator_3 = BusNavigator(bus_line_414)
+    bus3.add_bus_navigator(bus_navigator_3)
+
     bus4 = Bus(BUS4, BUS4_PASSWD)
-    bus3.add_line(bus_line_414)
+    bus_navigator_4 = BusNavigator(bus_line_414)
+    bus4.add_bus_navigator(bus_navigator_4)
+
     director.start()
     time.sleep(5)
     bus1.start()
